add <form> and 'submit' binding to submit on enter keypress
According to Knockout.JS documentatation, adding the 'submit' binding to the <form> tag, allows you to submit a form from the textbox by pressing the enter key. This is an improvement over having to navigate the cursor to the Send button.

Also, it seems that, to disable the typical <form> behavior, the 'submit'-binding function should .